replaced structure in order to try a better multimode structure
