+++ build.gradle
@@ -1,20 +1,23 @@
-group 'com.github.mydogtom'
-version '0.1'
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-apply plugin: 'java'
+buildscript {
+    repositories {
+        jcenter()
+    }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:2.2.2'
 
-sourceCompatibility = 1.8
-
-repositories {
-    jcenter()
+        // NOTE: Do not place your application dependencies here; they belong
+        // in the individual module build.gradle files
+    }
 }
 
-def RXJAVA_VERSION = '1.2.1'
+allprojects {
+    repositories {
+        jcenter()
+    }
+}
 
-dependencies {
-    compile "io.reactivex:rxjava:$RXJAVA_VERSION"
-
-    testCompile 'junit:junit:4.12'
-    testCompile "io.reactivex:rxjava:$RXJAVA_VERSION"
-
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
