+++ settings.gradle
@@ -1,3 +1,1 @@
-rootProject.name = 'rxtestwrapper'
-
-
+include ':app', ':rxtestwrapper'
